all working on local db and server. work needed on getting full app to recognize user is signed in.
