+++ src/main/java/com/webdevgroup/sp2101webdevegroupserverjava/controllers/EventController.java
@@ -10,7 +10,9 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@CrossOrigin(origins = "*")
+//@CrossOrigin(origins = "*")
+@CrossOrigin(origins = "http://localhost:3000", allowCredentials = "true")
+
 public class EventController {
 
     private final EventService service;
