+++ src/main/java/com/webdevgroup/sp2101webdevegroupserverjava/services/UserService.java
@@ -1,7 +1,7 @@
 package com.webdevgroup.sp2101webdevegroupserverjava.services;
 
 import com.webdevgroup.sp2101webdevegroupserverjava.models.User;
-import com.webdevgroup.sp2101webdevegroupserverjava.repositories.UserRepository;
+import com.webdevgroup.sp2101webdevegroupserverjava.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,7 +17,9 @@ public class UserService {
         return repository.findUserById(id);
     }
     public User findUserByUserName(String username) {
-        return repository.findUserByUserName(username);
+//        User foundUser = repository.findUserByUserName(username);
+        User foundUser = repository.findByUsername(username);
+        return foundUser;
     }
 
     public List<User> findAllUsers() {
