+++ src/main/resources/application.properties
@@ -1,7 +1,12 @@
-spring.datasource.url=jdbc:mysql://bac275b785f43b:e3d68073@us-cdbr-east-03.cleardb.com/heroku_a6f102cb7f7ae1c?reconnect=true
+
+spring.datasource.url=jdbc:mysql://localhost:3306/webdevgroup_user_schema?serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=password
-spring.jpa.hibernate.ddl-auto=none
+spring.datasource.password=webdevpw
+spring.jpa.hibernate.ddl-auto=update
+
+#spring.datasource.url=jdbc:mysql://bac275b785f43b:e3d68073@us-cdbr-east-03.cleardb.com/heroku_a6f102cb7f7ae1c?reconnect=true
+#spring.jpa.hibernate.ddl-auto=none
+
 spring.jpa.show-sql=true
 spring.jpa.hibernate.naming-strategy=org.hibernate.cfg.ImprovedNamingStrategy
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
@@ -9,4 +14,4 @@ spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.use_sql_comments=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.type=trace
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+
